tech debt moved delete actions to manage controller, moved tests as well.
