+++ src/SFA.DAS.Reservations.Web/Controllers/ReservationsController.cs
@@ -378,120 +378,6 @@ namespace SFA.DAS.Reservations.Web.Controllers
 
             return View("../EmployerReservations/FundingRestrictionNotification", viewModel);
         }
-
-        [Route("{ukPrn}/reservations/{id}/delete", Name = RouteNames.ProviderDelete)]
-        [Route("accounts/{employerAccountId}/reservations/{id}/delete", Name = RouteNames.EmployerDelete)]
-        public async Task<IActionResult> Delete(ReservationsRouteModel routeModel)
-        {
-            var isProvider = routeModel.UkPrn.HasValue;
-            try
-            {
-                if (!routeModel.Id.HasValue)
-                {
-                    _logger.LogInformation($"Reservation ID must be in URL, parameter [{nameof(routeModel.Id)}]");
-                    var manageRoute = isProvider ? RouteNames.ProviderManage : RouteNames.EmployerManage;
-                    return RedirectToRoute(manageRoute, routeModel);
-                }
-
-                var query = new GetReservationQuery
-                {
-                    Id = routeModel.Id.Value,
-                    UkPrn = routeModel.UkPrn.GetValueOrDefault()
-                };
-                var queryResult = await _mediator.Send(query);
-
-                var viewName = isProvider ? ViewNames.ProviderDelete : ViewNames.EmployerDelete;
-
-                return View(viewName, new DeleteViewModel(queryResult));
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "Error preparing for the delete view");
-                throw;
-            }
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        [Route("{ukPrn}/reservations/{id}/delete", Name = RouteNames.ProviderDelete)]
-        [Route("accounts/{employerAccountId}/reservations/{id}/delete", Name = RouteNames.EmployerDelete)]
-        public async Task<IActionResult> PostDelete(ReservationsRouteModel routeModel, DeleteViewModel viewModel)
-        {
-            var isProvider = routeModel.UkPrn.HasValue;
-            var deleteViewName = isProvider ? ViewNames.ProviderDelete : ViewNames.EmployerDelete;
-            try
-            {
-                if (!ModelState.IsValid)
-                {
-                    return View(deleteViewName, viewModel);
-                }
-
-                if (viewModel.Delete.HasValue && !viewModel.Delete.Value ||
-                    !routeModel.Id.HasValue)
-                {
-                    var manageRoute = isProvider ? RouteNames.ProviderManage : RouteNames.EmployerManage;
-                    return RedirectToRoute(manageRoute, routeModel);
-                }
-                
-                await _mediator.Send(new DeleteReservationCommand{ReservationId = routeModel.Id.Value});
-
-                var completedRoute = isProvider ? RouteNames.ProviderDeleteCompleted : RouteNames.EmployerDeleteCompleted;
-                return RedirectToRoute(completedRoute, routeModel);
-            }
-            catch (ValidationException ex)
-            {
-                _logger.LogInformation(ex, $"Validation error trying to delete reservation [{routeModel.Id}]");
-                return View(deleteViewName, viewModel);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, $"Error trying to delete reservation [{routeModel.Id}]");
-                var errorRoute = isProvider ? RouteNames.ProviderError : RouteNames.EmployerError;
-                return RedirectToRoute(errorRoute, routeModel);
-            }
-        }
-
-        [Route("{ukPrn}/reservations/{id}/delete-completed", Name = RouteNames.ProviderDeleteCompleted)]
-        [Route("accounts/{employerAccountId}/reservations/{id}/delete-completed", Name = RouteNames.EmployerDeleteCompleted)]
-        public IActionResult DeleteCompleted(ReservationsRouteModel routeModel)
-        {
-            var viewName = routeModel.UkPrn.HasValue ? ViewNames.ProviderDeleteCompleted : ViewNames.EmployerDeleteCompleted;
-
-            return View(viewName);
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        [Route("{ukPrn}/reservations/{id}/delete-completed", Name = RouteNames.ProviderDeleteCompleted)]
-        [Route("accounts/{employerAccountId}/reservations/{id}/delete-completed", Name = RouteNames.EmployerDeleteCompleted)]
-        public IActionResult PostDeleteCompleted(ReservationsRouteModel routeModel, DeleteCompletedViewModel viewModel)
-        {
-            var isProvider = routeModel.UkPrn.HasValue;
-            var deleteCompletedViewName = isProvider ? ViewNames.ProviderDeleteCompleted : ViewNames.EmployerDeleteCompleted;
-            var manageRouteName = isProvider ? RouteNames.ProviderManage : RouteNames.EmployerManage;
-            var dashboardUrl = isProvider ? 
-                _configuration.DashboardUrl : 
-                _urlHelper.GenerateUrl(
-                    new UrlParameters
-                    {
-                        Id = routeModel.EmployerAccountId,
-                        Controller = "teams",
-                        Folder = "accounts",
-                        SubDomain = "accounts"
-                    });
-            
-            if (!ModelState.IsValid)
-            {
-                return View(deleteCompletedViewName, viewModel);
-            }
-
-            if (viewModel.Manage.HasValue && viewModel.Manage.Value)
-            {
-                return RedirectToRoute(manageRouteName);
-            }
-
-            return Redirect(dashboardUrl);
-        }
         
         private async Task<ApprenticeshipTrainingViewModel> BuildApprenticeshipTrainingViewModel(
             bool isProvider,
